Refactor computer to use BigInteger, allow access to memory outside program area, and add in relative base support, and Day 9 both parts

+++ 2019-Kotlin/src/uk/co/cemerson/aoc/Day2/GetFinalProgramStateOutputConsumer.kt
@@ -1,16 +1,17 @@
 package uk.co.cemerson.aoc.Day2
 
 import uk.co.cemerson.aoc.Util.IntCode.OutputConsumer
+import java.math.BigInteger
 
 class GetFinalProgramStateOutputConsumer : OutputConsumer {
-    private var output: Int = 0
+    private var output: BigInteger = 0.toBigInteger()
 
-    override fun consumeOutput(output: Int) {
+    override fun consumeOutput(output: BigInteger) {
     }
 
-    override fun consumeFinalValueInPositionZero(output: Int) {
+    override fun consumeFinalValueInPositionZero(output: BigInteger) {
         this.output = output
     }
 
-    fun getFinalOutput(): Int = output
+    fun getFinalOutput(): BigInteger = output
 }
