+++ 2019-Kotlin/src/uk/co/cemerson/aoc/Util/IntCode/OutputConsumer.kt
@@ -1,6 +1,8 @@
 package uk.co.cemerson.aoc.Util.IntCode
 
+import java.math.BigInteger
+
 interface OutputConsumer {
-    fun consumeOutput(output: Int)
-    fun consumeFinalValueInPositionZero(output: Int)
+    fun consumeOutput(output: BigInteger)
+    fun consumeFinalValueInPositionZero(output: BigInteger)
 }
