+++ 2019-Kotlin/src/uk/co/cemerson/aoc/Util/IntCode/PrintLnOutputConsumer.kt
@@ -1,10 +1,12 @@
 package uk.co.cemerson.aoc.Util.IntCode
 
+import java.math.BigInteger
+
 class PrintLnOutputConsumer : OutputConsumer {
-    override fun consumeOutput(output: Int) {
+    override fun consumeOutput(output: BigInteger) {
         println(output)
     }
 
-    override fun consumeFinalValueInPositionZero(output: Int) {
+    override fun consumeFinalValueInPositionZero(output: BigInteger) {
     }
 }
