+++ 2019-Kotlin/src/uk/co/cemerson/aoc/Util/IntCode/SpecifiedInputProvider.kt
@@ -1,10 +1,12 @@
 package uk.co.cemerson.aoc.Util.IntCode
 
+import java.math.BigInteger
+
 class SpecifiedInputProvider(private var input: List<Int>) : InputProvider {
-    override fun getInput(): Int {
+    override fun getInput(): BigInteger {
         val input = this.input.first()
         this.input = this.input.drop(1)
 
-        return input
+        return input.toBigInteger()
     }
 }
